Break down `ActionButtonViz` into files
