+++ frontend/src/metabase/writeback/components/ActionButtonViz/ActionButtonViz.tsx
@@ -1,13 +1,7 @@
-import React, { useCallback, useMemo } from "react";
 import { t } from "ttag";
-import cx from "classnames";
+import ActionButton from "./ActionButton";
 
-import Button from "metabase/core/components/Button";
-
-import { DashboardWithCards } from "metabase-types/types/Dashboard";
-import { VisualizationProps } from "metabase-types/types/Visualization";
-
-const ACTIONS_VIZ_DEFINITION = {
+export default Object.assign(ActionButton, {
   uiName: t`Action button`,
   identifier: "action-button",
   iconName: "play",
@@ -51,60 +45,4 @@ const ACTIONS_VIZ_DEFINITION = {
       },
     },
   },
-};
-
-interface ActionButtonVizProps extends VisualizationProps {
-  dashboard: DashboardWithCards;
-}
-
-function ActionButtonViz({
-  isSettings,
-  settings,
-  getExtraDataForClick,
-  onVisualizationClick,
-}: ActionButtonVizProps) {
-  const label = settings["button.label"];
-  const variant = settings["button.variant"];
-
-  const variantProps: any = {};
-  if (variant !== "default") {
-    variantProps[variant] = true;
-  }
-
-  const clicked = useMemo(
-    () => ({
-      settings,
-    }),
-    [settings],
-  );
-
-  const extraData = useMemo(
-    () => getExtraDataForClick?.(clicked),
-    [clicked, getExtraDataForClick],
-  );
-
-  const onClick = useCallback(
-    (e: React.MouseEvent) => {
-      onVisualizationClick({
-        ...clicked,
-        extraData,
-        element: e.currentTarget as HTMLElement,
-      });
-    },
-    [clicked, extraData, onVisualizationClick],
-  );
-
-  return (
-    <Button
-      className={cx({
-        "full-height": !isSettings,
-      })}
-      onClick={onClick}
-      {...variantProps}
-    >
-      {label}
-    </Button>
-  );
-}
-
-export default Object.assign(ActionButtonViz, ACTIONS_VIZ_DEFINITION);
+});
